+++ libraries/helpers.rb
@@ -1,18 +1,6 @@
 module Nginx
   module Cookbook
     module Helpers
-      # pidfile is hard to determine on Debian systems.
-      # Upstream packages and older distro releases use '/var/run/nginx.pid'
-      # systemd based distros and Ubuntu 14.04 use '/run/nginx.pid' for their
-      # packages
-      def pidfile_location
-        if (node['nginx']['repo_source'].nil? || %w(distro passenger).include?(node['nginx']['repo_source'])) && (node['init_package'] == 'systemd' || node['platform_version'].to_f == 14.04)
-          '/run/nginx.pid'
-        else
-          '/var/run/nginx.pid'
-        end
-      end
-
       def nginx_binary
         '/usr/sbin/nginx'
       end
@@ -21,10 +9,12 @@ module Nginx
         case node['platform_family']
         when 'amazon', 'fedora', 'rhel'
           case node['platform']
-          when 'amazon', 'fedora'
+          when 'amazon'
             'https://nginx.org/packages/rhel/7/$basearch'
           when 'centos'
             "https://nginx.org/packages/centos/#{node['platform_version'].to_i}/$basearch"
+          when 'fedora'
+            'https://nginx.org/packages/rhel/8/$basearch'
           else
             "https://nginx.org/packages/rhel/#{node['platform_version'].to_i}/$basearch"
           end
@@ -51,6 +41,10 @@ module Nginx
         platform_family?('debian') ? 'www-data' : 'nginx'
       end
 
+      def nginx_group
+        platform_family?('debian') ? 'www-data' : 'nginx'
+      end
+
       def nginx_pid_file
         '/run/nginx.pid'
       end
@@ -59,16 +53,40 @@ module Nginx
         '/usr/sbin'
       end
 
+      def nginx_config_file
+        "#{nginx_dir}/nginx.conf"
+      end
+
+      def nginx_config_site_dir
+        "#{nginx_dir}/conf.http.d"
+      end
+
+      def nginx_default_packages
+        case node['platform_family']
+        when 'rhel', 'fedora'
+          %w(nginx)
+        when 'debian'
+          %w(nginx-full)
+        else
+          %w(nginx)
+        end
+      end
+
+      def default_nginx_service_name
+        'nginx'
+      end
+
       def default_root
-        '/var/www/nginx-default'
-      end
-
-      def nginx_site_enabled?(site_name)
-        ::File.symlink?("#{nginx_dir}/sites-enabled/#{site_name}") || ::File.symlink?("#{nginx_dir}/sites-enabled/000-#{site_name}")
-      end
-
-      def nginx_site_available?(site_name)
-        ::File.exist?("#{nginx_dir}/sites-available/#{site_name}")
+        case node['platform_family']
+        when 'rhel', 'fedora', 'amazon'
+          '/usr/share/nginx/html'
+        when 'debian'
+          '/var/www/html'
+        when 'suse'
+          '/srv/www/htdocs'
+        else
+          raise "default_root: Unsupported platform #{node['platform']}."
+        end
       end
 
       def debian_9?
@@ -78,32 +96,8 @@ module Nginx
       def ubuntu_18?
         platform?('ubuntu') && node['platform_version'].to_f == 18.04
       end
-
-      def passenger_packages
-        if platform_family?('debian')
-          packages = %w(ruby-dev libcurl4-gnutls-dev)
-          packages << if debian_9? || ubuntu_18?
-                        'libnginx-mod-http-passenger'
-                      else
-                        'passenger'
-                      end
-
-          packages
-        end
-      end
-
-      def passenger_conf_file
-        if platform_family?('debian')
-          if debian_9? || ubuntu_18?
-            ::File.join(nginx_dir, 'conf.d/mod-http-passenger.conf')
-          else
-            ::File.join(nginx_dir, 'conf.d/passenger.conf')
-          end
-        end
-      end
     end
   end
 end
 
-Chef::DSL::Recipe.include Nginx::Cookbook::Helpers
-Chef::Resource.include Nginx::Cookbook::Helpers
+Chef::DSL::Universal.include(Nginx::Cookbook::Helpers)