+++ spec/unit/resources/config_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe 'nginx_config' do
-  step_into :nginx_config, :nginx_install
+  step_into :nginx_config, :nginx_install, :nginx_site
   platform  'ubuntu'
 
   before do
@@ -34,6 +34,21 @@ describe 'nginx_config' do
         )
     end
 
-    it { expect(chef_run.template('/etc/nginx/nginx.conf')).to notify('service[nginx]').to(:reload).delayed }
+    it do
+      is_expected.to create_template('/etc/nginx/conf.http.d/default-site.conf')
+        .with_cookbook('nginx')
+        .with_source('default-site.erb')
+        .with_variables(
+          name: 'default-site',
+          nginx_log_dir: '/var/log/nginx',
+          port: '80',
+          server_name: 'Fauxhai',
+          default_root: '/var/www/html'
+        )
+    end
+
+    it { is_expected.to create_directory('/var/log/nginx').with_mode('0750').with_owner(nginx_user) }
+    it { is_expected.to create_directory('/etc/nginx/conf.d').with_mode('0750') }
+    it { is_expected.to create_directory('/etc/nginx/conf.http.d').with_mode('0750') }
   end
 end