+++ test/cookbooks/test/recipes/passenger.rb
@@ -1,3 +1,0 @@
-apt_update 'update' if platform_family?('debian')
-
-nginx_install 'passenger'