+++ test/integration/default/install_test.rb
@@ -1,0 +1,10 @@
+packages = case os.family
+           when 'debian'
+             'nginx-full'
+           else
+             'nginx'
+           end
+
+describe package(packages) do
+  it { should be_installed }
+end